backend/requirements.txt
channels==3.0.5
APScheduler==3.9.1.post1
GDAL==3.0.4
djoser==2.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djangochannelsrestframework==0.3.0
Django==3.2.10
django-filter==22.1
django-cors-headers==3.13.0
django-mail-templated==2.6.5
torch==1.12.1
torchvision==0.13.1
Pillow==9.0.1
matplotlib==3.4.3
asgiref==3.5.2
geopy==2.2.0
numpy==1.18.5
psycopg2==2.9.3
pytesseract==0.3.10
